notes and slight changes from review

+++ server/db/models/chocolate.js
@@ -6,7 +6,7 @@ const Chocolate = db.define('chocolate', {
     type: Sequelize.STRING,
     allowNull: false
   },
-  category: {
+  category: { // make category a model
     type: Sequelize.STRING,
     allowNull: false
   },
@@ -15,10 +15,10 @@ const Chocolate = db.define('chocolate', {
     allowNull: false
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER, // in cents OR Seq.DECIMAL
     allowNull: false
   },
-  stock: {
+  stock: { // default val, nonzero validation
     type: Sequelize.INTEGER,
     allowNull: false
   },
@@ -27,5 +27,8 @@ const Chocolate = db.define('chocolate', {
     defaultValue: 'https://www.google.com/search?q=kilwins+chocolate&source=lnms&tbm=isch&sa=X&ved=0ahUKEwi9sITYw6LWAhWq5IMKHcOjB3UQ_AUIDCgD&biw=1776&bih=845#imgrc=yGY-WoNoYQEVVM:'
   }
 })
+
+// instance method for increasing/decrementing stock
+// getter method convert price to dollars from cents
 
 module.exports = Chocolate
